Implement SEDMM
A full implementation of SEDMM's specifications.

SEDMM is divided into two parts, one that is Traits which are objects that bind some destruction behavior to a value, and the other part is Scopes, which are collections of state objects that can be Kill()ed. The killing process works by calling the defined behavior of a state object's trait, which cleans the value.